+++ src/clj/rems/email/core.clj
@@ -5,10 +5,13 @@
             [clojure.tools.logging :as log]
             [mount.core :as mount]
             [postal.core :as postal]
+            [rems.api.services.todos :as todos]
+            [rems.api.services.workflow :as workflow]
             [rems.application.model]
             [rems.config :refer [env]]
             [rems.db.applications :as applications]
             [rems.db.outbox :as outbox]
+            [rems.db.user-settings :as user-settings]
             [rems.db.users :as users]
             [rems.email.template :as template]
             [rems.scheduler :as scheduler])
@@ -20,12 +23,23 @@
     (template/event-to-emails (rems.application.model/enrich-event event users/get-user (constantly nil))
                               (applications/get-unrestricted-application app-id))))
 
-(defn generate-emails! [new-events]
+(defn- enqueue-email! [email]
+  (outbox/put! {:outbox/type :email
+                :outbox/email email
+                :outbox/deadline (-> (time/now) (.plus ^Period (:email-retry-period env)))}))
+
+(defn generate-event-emails! [new-events]
   (doseq [event new-events
           email (event-to-emails event)]
-    (outbox/put! {:outbox/type :email
-                  :outbox/email email
-                  :outbox/deadline (-> (time/now) (.plus ^Period (:email-retry-period env)))})))
+    (enqueue-email! email)))
+
+(defn generate-handler-reminder-emails! []
+  (doseq [email (->> (workflow/get-handlers)
+                     (map (fn [handler]
+                            (let [lang (:language (user-settings/get-user-settings (:userid handler)))]
+                              (template/handler-reminder-email lang handler (todos/get-todos (:userid handler))))))
+                     (remove nil?))]
+    (enqueue-email! email)))
 
 ;;; Email poller