+++ src/cljs/rems/application_list.cljs
@@ -72,10 +72,15 @@
            :description {:value (:application/description app)
                          :td [:td.description (format-description app)]}
            :resource {:value (format-catalogue-items app)}
-           :applicant
-           (let [applicant (application-util/get-applicant-name app)]
-             {:value applicant
-              :td [:td.applicant (format-applicant applicant)]})
+           :applicant (let [applicant (application-util/get-applicant-name app)]
+                        {:value applicant
+                         :td [:td.applicant (format-applicant applicant)]})
+           :handlers (let [handlers (->> (get-in app [:application/workflow :workflow.dynamic/handlers])
+                                         (filter :handler/active?)
+                                         (map application-util/get-member-name)
+                                         (sort)
+                                         (str/join ", "))]
+                       {:value handlers})
            :state (let [value (localize-state (:application/state app))]
                     {:value value
                      :td [:td.state
@@ -84,16 +89,21 @@
                           value]})
            :todo (let [value (localize-todo (:application/todo app))]
                    {:value value
-                    :td [:td.state
-                         {:class (when (current-user-needs-to-do-something? app)
-                                   "text-highlight")}
+                    :td [:td.todo
+                         {:class (str (when (current-user-needs-to-do-something? app)
+                                        "text-highlight ")
+                                      (when (:application/past-deadline app)
+                                        "text-danger"))}
                          value]})
            :created (let [value (:application/created app)]
                       {:value value
                        :display-value (localize-time value)})
            :submitted (let [value (:application/first-submitted app)]
                         {:value value
-                         :display-value (localize-time value)})
+                         :td [:td.submitted
+                              {:class (when (:application/past-deadline app)
+                                        "text-highlight text-danger")}
+                              (localize-time value)]})
            :last-activity (let [value (:application/last-activity app)]
                             {:value value
                              :display-value (localize-time value)})
@@ -112,6 +122,8 @@
                       :title (text :t.applications/resource)}
                      {:key :applicant
                       :title (text :t.applications/applicant)}
+                     {:key :handlers
+                      :title (text :t.applications/handlers)}
                      {:key :state
                       :title (text :t.applications/state)}
                      {:key :todo
@@ -135,7 +147,7 @@
 (defn- application-list-defaults []
   (let [config @(rf/subscribe [:rems.config/config])
         id-column (get config :application-id-column :id)]
-    {:visible-columns #{id-column :description :resource :applicant :state :todo :created :submitted :last-activity :view}
+    {:visible-columns #{id-column :description :resource :applicant :handlers :state :todo :created :submitted :last-activity :view}
      :default-sort-column :created
      :default-sort-order :desc}))