+++ CHANGELOG.md
@@ -17,6 +17,11 @@ Changes since v2.6
   New items will be created that use the new form while the old items
   are disabled and archived. The name of the new item will be exactly
   the same as before. See #837
+- Applications can be exported as CSV in admin menu (#1857)
+- Added a configuration option for setting a maximum number of days for handling a new application (#1861)
+- Reminder email about open applications to all handlers. The emails can
+  be sent by calling `/api/email/send-handler-reminder` on a cron schedule
+  or similar. The API requires an API key. (#1611)
 
 ### Enhancements
 - Application search tips hidden behind question mark icon (#1767)
@@ -37,6 +42,7 @@ Changes since v2.6
 - POSTing entitlements to entitlement-target is now retried (#1784)
 - [Rejecter bot](docs/bots.md), which rejects applications where a member is blacklisted for a resource (#1771)
 - "Assign external id" command for setting the id of an application (#1858)
+- Display on the actions page the handlers who are handling an application (#1795)
 
 ## v2.6 "Kalevalantie" 2018-11-12